Add parts. Correct some semantic issues

+++ spring/src/main/java/es/ua/dlsi/grfia/im3ws/muret/repository/PartRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
  * @author drizo
  */
 @RepositoryRestResource
-public interface PartRepository extends CrudRepository<Part, Integer> {
+public interface PartRepository extends CrudRepository<Part, Long> {
 }
